+++ filter/wiris/subfilters/client.php
@@ -53,9 +53,13 @@ class filter_wiris_client extends \core_filters\text_filter {
      */
     public function filter($text, array $options = []) {
         global $PAGE;
+
+        // Get Moodle current language.
+        $lang = current_language();
+        
         // Add the Javascript Thir-party library to the page.
         $PAGE->requires->js(
-            new moodle_url('/filter/wiris/render/WIRISplugins.js?viewer=image&safeXml=true&async=true&element=%23page&ignored_containers=[data-fieldtype="editor"]')
+            new moodle_url('/filter/wiris/render/WIRISplugins.js?viewer=image&lang=' . $lang . '&safeXml=true&async=true&element=%23page&ignored_containers=[data-fieldtype="editor"]')
         );
 
         // Uses the option 'safeXml' to True to render directly from Safe MathML as stored on the database.