+++ filter/wiris/docs/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v8.10.1 - 20 oct. 2025
+
+- fix: Filter Wiris changing language on quizzes.
+
+## v8.10.0 - 08 oct. 2025
+
+- feat: Comaptibility with Moodle 5.1.
+- fix: Tests.
+
+## v8.9.2 - 10 jun. 2025
+
+- feat: Updated to the newest SDK version (v8.13.2).
+
 ## v8.9.1 - 22 apr. 2025
 
 - feat: Add compatibility with PHP 8.4.
@@ -135,4 +148,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix & upgrade TravisCI job configuration.
 - Update project documentation by improving the main `README` file.
 - Start using `CHANGES` file as changelog.
-